📷 Update snapshot test for StudyHeader
